Change Person to Abstract Class
